Add hook SpecialCheckUserGetLinksFromRow
Hook creates extension point to adjust or replace the per-revision
links generated within CheckUser.  This allows extensions to provide
the most specific links desirable for CheckUser data.

The default generated links are provided to the hook as an associative
array.  This was done to make adjusting, rather than replacing, less
than a total pain.

Bug: 60275
Change-Id: Ib7848a222b6421536653e9661e622785780ec8ef