Add populateCheckUserTable script
This is meant as a replacement for CheckUser install.php's population of
`cu_changes` table. It has been written as a LoggedUpdateMaintenance
script so that we can use DatabaseUpdater for doing this automatically
instead of using our own script.

The functionality has been retained and the code is almost the same
except for minor code style changes and using Maintenance methods.
It also now outputs the start and end rc_ids which will be migrated
to cu_changes at the start of migration.

I haven't yet made this be used by update.php: that will be done
in a follow-up patch.

Change-Id: I2dee9ab4a8b91f674817f4847c2e109e70fb1df3