CheckUser "contributions" link should be a red link for non-existent accounts
In Special:CheckUser, when a user had tried create and account while this was prevented by abusefilter, his or her "contributions" link was still blue instead of a red link (as if that user actually exists). The patch add takes advantage of the existing yet unsued css class .mw-anonuserlink, and turn the link to red.

Bug: T170507
Change-Id: I669affa176f6c5b0dbf61ac3ca4e77a4fb6eb3e5