Validate username input before constructing subpage links
The value passed to these functions, in this case, is the string
a user provides for a user name. If this string contains characters
that are unsafe for a title (such as _ at the end) it will result
in an exception if passed as-is to SpecialPage::getTitleFor().

Regression introduced in 64727c57c7c279f88b.

Fix by validating it first via a Title::makeTitleSafe() check.

Bug: T217713
Change-Id: Ic488b7ee20757a2c0da06dbca1f4f8935033f059