Lookup users by their ID not their username
As pointed out by DannyS712 on the task, the code takes all the
necessary steps to determine if a user exists and what their
user ID is, then falls back to using the username itself for
looking up CU data. This not only causes T172477 but also results
in an inefficient query (there exists an index for cuc_user column
but none for cuc_user_text column).

Of note, the WWW part of the code in SpecialCheckUser.php does not
make this error; only the API version does.

Bug: T174277
Change-Id: I8b914af7f0f5c8a6962c621a0469dc4745f21c97