Fold 'caMultiLock' file into the 'ext.checkuser' module
This is a fairly small file only used on the same special page
that the other module is already loaded for. Merge these into
the same module so that subsequent views uses are faster for both
scenarios by not having a newly discovered file to (sometimes)
download. And, more importantly, this way reduces the overhead
in the startup module for all page views on all wikis.

Change-Id: I29e25002c21bc317bbc068a1a977d3a5185cc6d2