Display preliminary results in a table in Special:Investigate
This displays the results returned by the preliminary check service
in a table. It does so using a TablePager, overriding doQuery in
order to use the service to paginate on two columns, which are
unique in combination.

After this commit, the paging links have URLs containing the
necessary information for paging, but they do not yet work, since we
may not want to expose investigation data in the URL. This will be
solved in T239680.

Bug: T237298
Change-Id: Ib56ac5f2241c7bd2611c692769ffaea746a90b08