Paginate Special:Investigate results without exposing data in the URL
When requesting a particular page of results for the preliminary
check, make sure sensitive data (i.e. any user names and IP addresses
under investigation) are not exposed in the URL.

Do this by encrypting any sensitive data and passing it via the URL
encoded in a JSON web token, which expires after 24 hours.

Bug: T239680
Change-Id: Iba5ff88d203b55f2377e1f3b5b2e05f1f33af7d4