Reorganise preliminary check service for paginated results
The preliminary check tab on Special:Investigate will display basic
information about accounts for a given list of users across wikis.
The information includes number of edits, user groups and whether
the account is blocked.

If CentralAuth is available:
* Query the localuser table for all the wikis on which users from a
  given list are registered. Users may have accounts on a large
  number of wikis, so paginate on the user and wiki columns.
* For each row on the page, query the wiki's database for more
  information about the user.

If CentralAuth is not available:
* Query the local wiki's database for all the required information
  about a list of users.
* No pagination currently occurs. If investigations are likely to
  involve more than many tens of users, we should paginate.
* For code simplicity, this is done in several queries instead of
  one. We may want to change this.

For front-end display, we will need a TablePager that can paginate
on two columns that are unique in combination.

Bug: T238827
Change-Id: I03524cefdd631ebb2a3f2b5602788c0fc0485506