Update preliminary check classes to use multi-column pagination
Now that the IndexPager supports pagination on multiple columns,
pagination no longer needs to be done in the PreliminaryCheckService.

Remove handling of pagination from the service, and update the pager
accordingly. A downside of doing this is that the pager and service
are now more tightly coupled, meaning that any future creation of an
API will be more complicated. However, this is preferable to
duplicating the pagination handling, now that it exists in the
IndexPager.

This coupling means that PreliminaryCheckPager now needs to be aware
of whether CentralAuth is available, to select the relevant database
and index field.

Bug: T244517
Change-Id: I28aecc6a9117362351e84daea4ca4ee75117e7a7