Add highligting on hover to Special:Investigate results tables
This adds highlighting to the results tables for the preliminary check
and compare analyses.

This allows users to see matching values in "columns of interest".
These columns are:
* registration date (preliminary check)
* wiki (preliminary check)
* IP address (compare)
* user agent (compare)

= Hovering =

On hovering on a regular table cell, the whole row gains a pale
highlight. On hovering on a table cell in a "column of interest", that
cell gains a dark highlight, and the rest of its row gains a pale
highlight. Additionally, any cells in the same column with matching
data also gain a dark highlight, and their rows gain a pale highlight.
These highlights disappear on mouseout.

= Pinning =

For table cells in a "column of interest", a button for pinning the
highlight appears on hover. If clicked, any highlights caused by
hovering on that cell remain after mouseout. After mouseout, the
button remains visible, and can be clicked again to unpin the
highlights.

Once a highlight is pinned, all other highlighting (including on
hover) is disabled until the highlight is unpinned. All pin buttons in
the pinned column toggle on, and any can be clicked to remove the
highlight.

Bug: T237299
Change-Id: I80b6c44ad7bd083271c2b00780cba6d53518e3fa