Improve performance of query for Special:Investigate Compare tab
Before this, the query for the Compare investigation could access an
unlimited number of rows, which are then grouped by user, IP and user
agent. Accessing an unlimited number of rows is far from ideal; even
worse is sorting and grouping an unlimited number of rows.

This restructures the query so that a limited subquery is performed
for each target. This imposes limits while ensuring that we get at
least some results for each target. A total limit of 100k rows is set
by default (see task for more details), and this is shared equally
among the targets.

If the limit is hit for any targets, show a warning explaining which
targets have incomplete data.

Bug: T245499
Change-Id: Iecad40c9d6c56393a6e92315b1b6bea116c5d749