Fix SQLite support with improved query on Special:Investigate
Some databases, like SQLite, do no support an ORDER BY or LIMIT with a UNION.
Thankfully, MediaWiki provides this information on the IDatabase interface.
If the database does not support this feature, then we'll remove these options
from the query. This means the query will be unbounded, but there isn't a way
to prevent this.

Bug: T246917
Change-Id: Iade6c7170c27c74a4c3422cb87f150485435a4f3