CompareService: Select row ID in subquery
CompareService::getQueryInfo builds a query involving a subquery for
each target, creating a union table in which duplicate rows are
filtered out. The filtering is necessary because a user and an IP used
by that user could both be targets, in which case the subqueries will
find the same row.

Different rows in cu_changes can have the same user, IP, user agent
and timestamp. Before this commit, multiple such rows appear to be
duplicates in the union table, and are incorrectly filtered out. After
this commit, the row ID is also selected, to ensure only true
duplicates are filtered.

Bug: T245499
Change-Id: If7c9090485c1d166d4ce9bd53fd65c7eb8d4253c