Add buttons to Special:Investigate for adding extra IP targets
In the results table in the Compare tab, add buttons to cells
containing IP addresses. Clicking a button refreshes the page with
results for the same investigation, with this IP as an additional
target. If the IP in the cell is already a target, the button is
disabled.

See also f8ae40c6cbe87e, which does a similar thing for usernames.

Bug: T244816
Change-Id: I50e13757028dee4adaf941520a1a7c86fb987d3e