Adds data with basic formatting to the timeline tab
This patch introduces the timeline pager that takes care
of pagination and makes use of the query generated in TimelineService
to fetch the relevant data. It also introduces TimelineFormatter which
handles how cu_changes records are presented to the user.

Proper formatting of the different cu_changes entries will be done as
part of T247641.

Bug: T247640
Change-Id: I0a6f211452317f6cf392260bd8f6fe97b9d78b3b