Strengthen filter for Compare investigation on Special:Investigate
Before this commit, filtering user names and IPs simply removes the
filtered item from the list of targets. E.g. if a result row contains
a user name that is a target, but an IP that is filtered out, the row
will still be included.

After this, no results are included for a filtered user name or IP.
E.g. if a result row contains a user name that is a target, but an IP
that is filtered out, the row will be filtered out.

Bug: T247281
Change-Id: Ia9fb236ad160ffaac15f330031c878d849ebf4dd