Move token-query business logic into its own class
The goal is to handle all token related business logic
in a central place that could be accessed by the special page and
the different pagers.

This work is paving the way for the TimelinePager that will not
inherit from a TablePager and thus the need to move this functionality
out of InvestigatePager.

It also means that we are getting rid of InvestigatePager since it
will not be needed anymore.

Bug: T248996
Change-Id: I482834efa3e77c9ca3a887c286f8ca9e74a374d8