Show tool links underneath IP addresses
Add menu options for the tool links found in the message
'checkuser-toollinks'.

Tool links may differ between wikis; the list of links to include
is currently defined in the 'checkuser-toollinks' message. The
format is not standardized across translations, so there may be
different separators, etc.

This parses the translated message and extracts the href attribute
and text for each link. It builds the menu options from these.

Since the JS parser isn't full-featured, the message is parsed in
PHP.

Bug: T250290
Change-Id: I9c50a4f5bcc33f340fcab6bf8bb6b8fa99a8ea82