Trim comment field content prior to insertion
CheckUser populates the cuc_actiontext, cuc_agent, cuc_xff and cuc_comment
from various UGC, such as MediaWiki system messages and request headers. These
fields are all defined as VARCHAR(255) binary in the DB schema, so they can't
fit more than 255 bytes.

On installs that use MySQL/MariaDB in strict mode, this means that an action
with a too long comment or too long User-Agent header will cause the INSERT
into the checkuser table to fail. Worse still, this also prevents a recent
changes row from being created for the action (e.g. edit) that triggered the
INSERT!

As a fix, update the logic to truncate these fields before attempting to
INSERT them; on non-strict mode systems, this should be the same as the
current behavior, while on strict mode systems, this should prevent the error.
This fix also does not require a schema change or other migration in WMF
production.

A limited version of this fix has been applied to 1.19 CheckUser with strict
mode MySQL 5.7.25.[1][2]

Also swap out some trivial wfGetDB() calls to getConnectionRef(), as this
patch introduces use of the MediaWiki service container.

---
[1] https://github.com/Wikia/app/commit/62c9079665911ea9d561f90f
[2] https://github.com/Wikia/app/commit/31f511b319b4259a41c1bd71

Bug: T199323
Change-Id: I33b18764d49abf813898fc5006673f6b1b87712b