Add feature for copying Special:Investigate results table
Provide an interface for copying the results table in the Compare
tab. Use the mw.widgets.CopyTextLayout for copying the text. This
widget does not have a PHP implementation.

The wikitext is generated by Parsoid from the HTML, which is
sanitized to remove unnecessary classes and data.

The feature is only available for the Compare tab, and when
JavaScript and Parsoid are available.

Bug: T251361
Change-Id: I596701be6d727ec2379b7ff747d11164312f74fb