Fix a test on SQLite
This is now using a conditional which checks for the DB being used to
run tests. Ideally this should check both scenarios, but doing so is not
trival at the moment, because you have to mock the DB object (and keep
it able to return some data), mock the LB object, etc.

Bug: T253235
Change-Id: I40eb0f72a54deb2aaa64c4746d72eb464ba6aba0