Successful login attemps should be logged for the user account
For unsuccessful login attempts, it is best to only associated the
entry with the IP and not the user (after all, chances are that
the attempt was made by a third party). For successful login
attempts, however, the entry in CU data should be associated with
the user and appear in a check on that username.

Bug: T253946
Change-Id: Ida034a20f0756e31488c6fe76c3e67b7a8ade40d