ComparePager: Make the results table sortable if there is only 1 page
Since the results can't be sorted with an indexed query due to the
way we aggregate them, they can't be sorted across pages. This means
we can't make use of the TablePager's sorting infrastructure.

Instead, use the jQuery tablesorter module, which handles table
sorting within one table. Only make this available if the results all
fit on one page.

Bug: T237300
Change-Id: I7a40a1aa14b9af6db269407ab91a894ca8533aa7