Add a new message for Special:Investigate tool links
Before this, Special:Investigate was using the old message, used by
Special:CheckUser.

However, the messages are used in different ways. The old message is
used to create a line of text, but the new message is only used to
create links (each with a label and a URL). That meant that using the
old message sometimes led to nonsensical output.

Bug: T255504
Change-Id: Ie3d376e8610b34038ee4bc87c1330ef4474357a8