Sort "get edits" consistently, no matter the number of results
When running "get edits" on a user account, CheckUser first checks
to see if the estimated row count is greater than the limit (by
default, 5000). If it is, it adds a "checkuser-limited" message to
the output, runs a slightly different query, and formats the output
in a slightly different way. This patch removes the last two
differences, leaving only the "checkuser-limited" error message.

The only difference in the queries is that, in the "over 5000 edits"
case, the edits were sorted first by IP address and then by
timestamp. The "under 5000" case sorts only by timestamp.
Lexicographic sorting of IP addresses does not provide any benefit
to CheckUsers.  The comment says that this is done to avoid a
filesort, but according to "explain", a filesort is done either way.
Note that the CheckUser API does not differentiate between these two
cases, and seems to have acceptable performance sorting by timestamp
alone, even for users with a large number of recent edits.  Since
this was the only difference between the two methods, this patch
deletes `userEditsRequestLimitExceededDB`, and calls
`doUserEditsDBRequest` instead.

Finally, since both cases are now calling the same two methods, this
patch removes the duplicate code from the end of
`doUserEditsRequest`. The only functional change here is that
`set_time_limit(60)` is now called for both cases, instead of only
the "under 5000" case.

Bug: T257538

Change-Id: I60ef9404d54c980ee4a45adf7b1ffeb0c8989314