Validate username input before constructing subpage links
The value passed to these functions, in this case, is the string
a user provides for a user name. If this string contains characters
that are unsafe for a title (such as _ at the end) it will result
in an exception if passed as-is to SpecialPage::getTitleFor().

Regression introduced in 64727c57c7c279f88b.

Fix by validating it first via a Title::makeTitleSafe() check.

Based on a275b06c0e8466c713256173ab4d8e909d5727bb,
which was reverted due to T265606.

Bug: T217713
Change-Id: Idafb946b464df1be1129e76c4413f9d10cb51097