Stop using ActorMigration for log_user
I joined on actor to select by user name, so that after the revision
table is fully migrated, we can remove the user object construction.

User object construction was introduced for ActorMigration, but before
that, it used User::idForName() to avoid a join. I think it's better to
use a join.

Bug: T278917
Change-Id: I9ec4312aae4b8b00354e1f990379c4f1fb841221