Change actor migration to write both
259 is the value of SCHEMA_COMPAT_READ_OLD | SCHEMA_COMPAT_WRITE_BOTH

I want to do this to prevent potential problems in wikis running with
the extension on the master version.

I would like to add a script in a next patch that will populate the
values in the cuc_actor column. This script will be added to the
onLoadExtensionSchemaUpdates hook. Since it is a LoggedUpdateMaintenance
script, it will only run once as long as update.php --force is not run
manually. Therefore we should set the migration setting to _write_both
before adding the script.

Bug: T233004
Depends-On: I49e7f066fcd35c725e3f052ffc4bc3aaf743eeb7
Change-Id: I0321a4c23cf4163e091f7a5b4c5be802ca0a8265