Show CUs whether the password matched on login attempts to locked accounts
Allow Checkusers to see whether the password was correct for login
attempts that fail. If the password is correct show this in the
CU entry alongside the pre-existing 'failed to login' message.

On any login attempts to locked accounts that provide the matching
password then set the user in the CU entry as the user that the
client tried to log into. This is because, if the account was not
locked, the request would have been otherwise successful.
This also means that attempts to log into a locked account
with the correct password also show when checking that locked
user.

This patch allows the CheckUser extension to read and use the
AuthenticationResponse::$failReasons. Currently only CentralAuth
implements provides this info, but any code that wants to notify
the Checkuser tool that the password was correct should use
"good password".

This change is compatible with older versions of MediaWiki
that do not support AuthenticationResponse::$failReasons
by keeping the existing behaviour simply saying 'failed
to log in'.

Bug: T303192
Depends-On: I7b2d9579a518a6c02f05281b1016e31e0d086fe7
Depends-On: I379d95500f89d26882e42502a07f80e145e8359c
Change-Id: Id69bfa223c2d670850492f42de9c5fd8111776f8