Allow CheckUserLog API to parse CIDR ranges
Ensure that the CheckUserLog API parses CIDR ranges in the same way
that Special:CheckUserLog does, such that checks on IPs inside the
CIDR range target are included in the result.

The Special:CheckUserLog behaviour seems to be the intended behaviour
based on the documentation for the cultarget API parameter saying it
supports CIDR ranges.

Bug: T296899
Change-Id: Ibe409055392ece9b5c9db02c268bd9717accc7f2