Ensure that on unprovided XFF header the DB stores nothing
On sqlite a unprovided XFF header, which results in a false
value from the getHeader() function, is currently later stored
as the string value 0. Fixed by converting to the empty string
if the value returned by getHeader() is false as an empty XFF
header should result in the same result as no provided
XFF header.

Bug: T39613
Change-Id: I245c7af29dfd2fcd3210595acb350c77896d3c86