Check rights for Investigate GUI and InvestigateBlock permission errors
Hides the "Log" link from Special:Investigate when
the user does not have the log right. Also hides
the block form from Special:Investigate when the
user does not have the block right. If the user
tries to acccess Special:InvestigateBlock when they
have the checkuser right but no block right show
them the PermissionsError message for lacking the
block right instead of the checkuser right as is
currently done.

Bug: T268152
Change-Id: Ibf73915d352a7fc03e5c60eb1790d1e8e970b610