Use OOUI in SpecialCheckUserLog
Make the Special:CheckUserLog form to use OOUI.
Also change the form so that you can search for
both the target and initiator, as well as allow
searching between two dates. The date selector
is now the OOUI date selector and also now allows
searching by day of the month in addition to the
month and year that was already possible.

Cherry-pick to master. Resolve conflicts.
Unmangle conflicted files. Thanks to TheresNoTime
for that.

Bug: T266586
Change-Id: Id66156118a151ce061b2b7c4a8c6a06dbbfb519c