Simplify SpecialInvestigate's addIndicators() parameters
Simplify SpecialInvestigate::addIndicators() parameters by
using one parameter to specify if the current page is a
subpage of Special:Investigate. Current behaviour only
has $newTab and $logOnly be different if the user is
on a subpage or not on a subpage. There is no seen
need to have this current behaviour changed, so
simplification is best.

Change-Id: I08f2b6b822737be8323928576898982e0eb3f838