Add tests for Special:CheckUser
Add tests for userWasBlocked() and
getTimeConds(). Also provide a function
in the test file to create a SpecialCheckUser
object for testing on. Split from an abandoned
patchset that was a little too large. It's at 799015.

These are kind of half unit test and integration
tests. They test individual components of
SpecialCheckUser.php.

Bug: T175920
Change-Id: Ic798f84a2e0b44969c6be67cf1ccf52a6b8c0542