Allow searching CheckUserLog by check reason with wildcards
Add to Special:CheckUserLog the ability to search the reason
for the check directly or through wildcards on either ends
of the search string. Wildcard searching can be globally
disabled with a global variable. By default wildcard search
is disabled, though this may change after it's performance
is evaluated on WMF wikis. If enabled, a checkbox appears
that allows the search to include wildcards which is left
unchecked by default.

This patch also migrates the cul_reason to the comment table
and adds a new row that stores a foreign key to the comment
table for the plaintext version of the reason. The plaintext
version of the reason is the reason stripped of any wikitext
that is parsed when showing the reason in the CheckUserLog
(such as wikilinks but not template uses). The plaintext
value is stored in the DB so that those searching to not
need to include wikitext that isn't shown in the log
when viewing entries. If not included this would make
searching harder. When searching both the plaintext
and non-plaintext version of the reason is checked
against the search string.

This patch does not allow this functionality through the
API, which will be added in a later patch.

Bug: T16699
Change-Id: Ie257444bcb68ce1b2bdfbaf9235a0a651d21788e