Show users under investigation even without the block right
Ensure that the users under investigation are still shown
even if the user does not have the block right. This patch
changes the block form such that it still shows but does
not add the button that links to the block page if the user
lacks the block right. In a previous patch a user without
the block right would not see a list of those they were
investigating.

Bug: T268152
Change-Id: I5a0fc67ef9675a711a6d71e1f79b0828e56f21ea