Add User:Amalthea/culoghelper.js into the extension
Link the timestamp in the log if searching for initiator, target or reason.
The link is to the CheckUserLog without the current search parameters but
with an offset of an hour after the timestamp. Included in the link
is a query parameter named 'highlight' that is used to highlight the log
entry on the log page that is linked to. If JavaScript is enabled on the
client browser then the page will also scroll down to the log entry that
was highlighted. The highlight parameter can also be included in the
query parameters of any log link.

This adds into the extension the features provided by the userscript
https://en.wikipedia.org/w/index.php?title=User:Amalthea/culoghelper.js
&oldid=805789812
This patch is inspired by that userscript, but is completely re-written in
PHP and a small amount of JavaScript. JavaScript is only needed to make
the browser scroll to the highlighted entry, so if the client has disabled
JavaScript then they will just need to scroll down themselves.

Bug: T309925
Change-Id: I1efd97098a9a2f41951ba7ac67125f516e78e5e4