Minor bug fix for OOUI conversion of Special:CheckUser
In converting Special:CheckUser to use OOUI a mistake
was made in the code that made the default be "Get users"
instead of "Get IP Addresses" for the radio options.
This wasn't noticed due to the similarities in the strings
'subuserips' and 'subipusers' when reviewing the code before
commit, and wasn't noticed during testing.

Bug: T309815
Change-Id: Ia9fbf178bd4c0adc3c9c3719c38d52cfd96223d0