Add cul_reason_id and cul_reason_plaintext_id into cu_log
Adds cul_reason_id and cul_reason_plaintext_id into the cu_log
table. Split from Ie257444bcb68ce1b2bdfbaf9235a0a651d21788e on
advice on IRC.

The columns will be used to store the reason for the check
in the comment table to avoid duplication in the DB. The
plaintext value is used to allow searching through the reason
which is intended to be added in a later patch.

Bug: T16699
Bug: T309982
Change-Id: I864191358f405daf5c7e24668effed671dc0af55