Use OOUI in the CIDR calculator and checkuserblock form
Use OOUI in the CIDR calculator JS form and the checkuserblock
form that is present on the "Get users" result page.
No functionality changes are made to the forms, with the only
changes being cosmetic (i.e. convering to use OOUI).

To make the conversion possible a helper class is used that
allows the SpecialCheckUser.php code to define it's own form
element so that the results on "Get edits" can be wrapped in a
form that includes the non-OOUI elements. The changes
to make the CIDR calculator use OOUI depend on a change in core
that allows an ID to be specified for the 'info' form field type.

Bug: T310019
Bug: T309815
Depends-On: If4724e6699aa3ec076817a7c4727cbff184b94e7
Change-Id: Ief7bfdfe06a4bce5c214f6754617c9b75c4b3965