Remove reverted message keys from qqq.json
Remove reverted message keys from qqq.json that were reverted out
when a patch was reverted. The keys are unused and should not be
used for the time being. Keeping this in qqq.json causes an error
on all patches that base on top of the latest translatewiki.net
patch.

Change-Id: I8ba7f8353c84c4991ba5692fdabea990c736166e