Make 'Get edits' selected if 'Get IP Addresses' is used with an IP
The OOUI FormFields use the default only if the value was not set in
the request, so ensure that the 'Get edits' option is default
selected if an IP is provided with a 'Get IP Addresses' request. This
restores the pre-OOUI conversion behaviour. Found and raised by
AntiCompositeNumber in phabricator (thanks for raising it).

Bug: T309815
Change-Id: I24e81542a0b4f6c38dec422d5f4762cb374e1145