Add required right tests for SpecialCheckUser and SpecialCheckUserLog
Add tests that check if the restriction for the Special:CheckUser
and Special:CheckUserLog page is defined such that the right is
containied in the checkuser group but not a user with no groups.
Also test that a new user group with only read and checkuser /
checkuser-log respectively has access to test that the restriction
is correct.

Bug: T175920
Change-Id: I3e43bf4e24aee54d566962fd4dc5e880096be1be