Show descriptive error on too small CIDR and disable result input
Show a more descriptive error on a too small CIDR such that '!' is
replaced with 'Common CIDR is too small' and shown in the input
result and the number of affected IPs is shown below in the usual
place. Also make the output CIDR input box be uneditable as it is
the result of a script. To do this without disabling selecting or
making the input greyed out, some new classes that override the
TextInput field are used.

Bug: T310425
Change-Id: I1b143126ff0905abe7b7c83dbca887063d244725