Use to and from in culdir api help message instead of start and end
Use culto and culfrom in the culdir API help message, as the
checkuserlog API uses the non-standard to and from (instead of
start and end). Aliasing to start and end would be better but in
discussion on IRC there seems to be no support in mediawiki code
for API parameter aliasing. If support is added doing that would
be preferrable.

Bug: T113205
Change-Id: Icee859ea6190e52c20606e4826793be0cc0ab6ac