Remove dead-link Tor node checker from checkuser-toollinks
Remove the Tor node checker from the checkuser-toollinks message
as it has been dead for several years. The task to maybe replace
it has been inactive for several years, though agreement on
removal was there at the time. Because no replacement has been
agreed on and this is a deadlink, just remove it without replacement
for the time being.

Bug: T190542
Change-Id: I149a40ddecaf4a95dd511f51e7fc26a390c6b2ad