Minor bug fix for timestamp linking function
Minor bug in LogPager::generateTimestampLink
had a check for a 'reason' search option that only exists in a
patch not yet merged (being able to search by reason in the CU log).
This raises a PHP warning.
Remove this check as the reason option is not included.

Bug: T309925
Follow-Up: I1efd97098a9a2f41951ba7ac67125f516e78e5e4
Change-Id: I729ad34ed403b79f101fe22f1ed31916f71a25dd