Link users in password reset rows in CheckUser
Link the user who had the password reset email sent to them in the
CheckUser row. The CU Hook that inserts this row is only called
once the code knows the user exists, so there is no need to store
the username without a link as the link will be for a existing
user account.

Depends on I5e3e9e047108ea67cfab8a730b45a98b057a7c1d because that
commit re-writes the hook that this commit modifies.

Change-Id: I56a6e2a8e59de6d04f827c238547fff7bd1f8f7c