Ensure the input for replacing the userpage has the 'blocktag' ID
A mistake in the OOUI conversion of the block form meant that the
ID for the block form input for replacing the userpage was changed
from 'blocktag' to 'tag'. This patch ensures that 'blocktag' is
used as this is used and dependended on by at least it.wiki's
MediaWiki:Group-checkuser.js.

Bug: T311041
Change-Id: I698cce3c2f583676cd6ed21d236a7fa59593282c