Remove Traceroute and RBLs as broken and fix WHOIS link for tool links
Remove the 'Traceroute' and 'RBLs' links from checkuser-toollinks
as these are deadlinks with no way to fix the link to use the same
site. Fix the WHOIS/RNDS link so that the link is no longer a 404.
This still uses the same site. Replacements will need to be
discussed, but as these are broken there is no need to keep them
around as they won't work anyway.

Bug: T311345
Bug: T311346
Bug: T311342
Change-Id: I748fd2e2eac07e61a0d82edd60dbc26e8c6b4f39