Make checktype options defined as constants for better readability
The string values 'subipusers', 'subedits' and 'subuserips' are
not very readable and also (except subedits) do not easily
correspond to the names of the check types. Using constants with names
that are the same as the en.json names for these subtypes will
make the code easier to read while still keeping backwards
compatability with tools that rely on the check types being these
string values.

Change-Id: I6cc3e991ccfe319f132ed17c5f20d79421808dab