Use FormOptions over $request->get...() for future simplification
Use FormOptions (which is used in Special:Watchlist and
Special:RecentChanges amongst others) so that the request
parameters are stored in the object properties. This means that
in future commits the request parameters do not need to be passed
around via parameters to functions if no custom validation /
modification occurs to them. Using an array is also possible but
I chose FormOptions as this handles edge cases and defaults nicer.

This will have the benefit of making the implementation of T311363
and T311375 not lead to many more parameters to multiple functions
which will reduce the code size. Also modified are the reason tests
to account for the reason now being inside the opts value.

Depends on the constants change as it's to interweaved too have
separate without a edit conflict.

Change-Id: Ica6fc7b90a9ac9a8176107d521e6c77654da9a39