Move to relying on $this->opts for period and checkuserblock form
Move to relying on $this->opts to get the period and to get the
form parameters for the checkuserblock form on 'Get edits'.
This does groundwork for moving to a method that gets the period
for T311363 and removes the need for the blockParams array to be
created except when the checkuserblock form is submitted. Also add
parameter types to the doMassUserBlockInternal() and tagPage() methods.
This also removes the unused $tag and $talkTag parameters from the
'Get users' request functions as getting the value set by the user
is now done by HTMLForm.

Change-Id: I5dd9bbca5042ec2c80924d92acb6f1f55b213eb9