Add a link to check the /64 for all IPv6 on 'Get IP Addresses'
Add a link to check the /64 for all IPv6 addresses on the 'Get IP
Addresses' results page. Blocking by the /64 for a IPv6 is commonly
described as for nearly all cases being similar to blocking a single
IPv4 address on enwiki. This is also the case for checking IPv6
addresses using CheckUser. The link makes it quicker to check the
/64.

Bug: T312994
Change-Id: I35ee8876f639922c72e3c524c08de14ce8c1f1c3