Change cu_log actor migration to write both
259 is the value of SCHEMA_COMPAT_READ_OLD | SCHEMA_COMPAT_WRITE_BOTH

I would like to add a script in a next patch that will populate the
values in the cuc_actor column. This script will be added to the
onLoadExtensionSchemaUpdates hook. Since it is a LoggedUpdateMaintenance
script, it will only run once as long as update.php --force is not run
manually. Therefore we should set the migration setting to _write_both
before adding the script, to prevent that there are rows in the db in
which cul_actor is NULL.

Bug: T233004
Depends-On: Id03b399f279d83ba510773e15f7009cbebb101bf
Change-Id: I3e6fc369eb97ad36ed589dca671bb7ff4c3369d6