Fix isValidRange and getIpConds tests and document test data
Fix the SpecialCheckUser's tests for isValidRange and getIpConds
as these relied on $this-lowerThanLimitIPv6 and the same for IPv4
being set when the data provider was callled. From my testing this
is not the case. I've moved these tests to a seperate test case so
that the properties will be set. Also document the test data for
the dataProviders for those tests and other tests in SpecialCheckUser.

Bug: T175920
Change-Id: Id179de70aa058dfdf75919b30d83a25f6e2f8550