Ignore broken code coverage reports for ServiceWiring.php
Ignore code coverage reports for ServiceWiring.php as the tests
that test the file cannot have an annotation for the entire file
as PHPUnit does not support providing file names in the @covers
annotation. This means that, due to MediaWiki's PHPUnit config
having test failures for missing @cover annotations the test file
only has the option to @coverNothing. This means that while the
ServiceWiring.php file is tested it cannot be marked as such.

Therefore ignore the PHPUnit code coverage report for the entire
return statement so that the code coverage report does not
incorrectly say that the file has no tests.

Change-Id: I7cd210788873de92a4e81ea019febf82e1f9ed8a