Further tests for Hooks.php
Reduce code duplication across the updateCheckUserData tests by
using a common test function that is called by each test method
that tests updateCheckUserData. Also fix the test for when no data
is to be saved as it previously didn't call the function it was
testing first. Also add tests for updateCUPasswordResetData.

Bug: T175920
Change-Id: I53d66c39bafa0d9579db3ca9e4b8993e7290dce8