Disable showing the paging navigation bar until paging for Get IPs works
Disable showing the paging navigation bar until the paging feature
for the Get IPs results works. Testing that I was doing just before
the merge found the issue and I've yet to get a patch that fixes it.
Disabling the functionality for the get IPs results will address the
issue for the time being.

Bug: T311375
Change-Id: I3472d8cbfa7e75616f6739852de52d6ce42720dd