Add a hook to allow modification of the cu_changes row before insertion
Add a hook that allows code that hooks on to it to change details
about the row to be inserted such as the request IP, XFF and
user agent. The hook is run after the request IP and XFF is got
from the request but before the code applies default values and
truncation for size limited fields.

Bug: T311340
Change-Id: I869626f744858a1478b1064e1e2b3534fd5af113