Refactor SpecialCheckUser.php checktype routing code
Refactor SpecialCheckUser.php so that the functions used for each
request subtype are removed and code related to it is moved into
the execute() method and de-duplicated where possible. Also move
the creation of new instances of the pagers into one function
to place them together so that if the services provided to the
pagers change in the future the code to change in SpecialCheckUser.php
is all in the same function. Finally also add a test for doesWrites.

Change-Id: Ibeb3903d381c28c6737482f3d596d002d6d2a30b