Add tests for the CheckUser pagers and CheckUserLogService
Add first tests for the CheckUser get users and get IPs pagers
and test the CheckUserLogService fully. The CheckUser pager tests
test some helper methods and also starts on tests for the formatRow
method by checking the provided arguments sent to the TemplateParser.
These unit tests won't test the mustache files but instead test the
parameters that are provided to them.

Bug: T175920
Depends-On: Iaac57d908c0617c201ad8f8ca500e93c66c653ce
Change-Id: Ibd8543789716ec8cfad2c316ed41f687f016a124