Start writing to the comment table for the cu_log reason
If the appropriate schema migration stage config allows it, make the
CheckUserLogService write to cul_reason_id and cul_reason_plaintext_id.
Also add some tests to verify that the new fields are only written to
if the migration stage allows it and also test that when the migration
stage allows writing to the new fields that the values written are
correct via some test cases.

The plaintext reason is the text that is shown when the reason is
displayed in the CheckUserLog. This means that the wikilink syntax is
removed but not the template or external URL syntax as URLs / templates
are not parsed when showing a log entry reason. It's intended use is to
allow searching in the log.

Bug: T309454
Depends-On: I6d3a50799808eb473d37988eedbf33a8d1dddbcd
Change-Id: Ie8beab83bb8a7eea2449aad84d53479af0fc2d5f